Lots of smaller changes; tried to find out why SDF shapes aren't properly clickable

+++ test/src/me/anno/tests/structures/PermutationsTest.kt
@@ -16,6 +16,6 @@ object PermutationsTest {
             ctr++
             assertTrue(all.add(it.toList()))
         }
-        assertEquals(base.size.factorial(), ctr)
+        assertEquals(base.size.factorial().toInt(), ctr)
     }
 }
